Bug 1874317 - Prune dead code from RevocableStore. r=nika
We aren't using the `empty()` method, which means we can remove both it
and the `count_` member, and avoid a benign data race that TSan caught.

The pointer back to the `RevocableStore` could in theory be replaced
with a boolean as a result, but this patch is limited to simple code
removal.

Differential Revision: https://phabricator.services.mozilla.com/D198483

UltraBlame original commit: 8dc6ba4f31e07ae650c52d4bd60dd477a753febf