Bug 1869876 - Parse the gecko profiler arguments properly for exra profiler run r=kshampur,perftest-reviewers
Previously we were only parsing the `gecko_profile_*` arguments for
`--gecko-profile` and not for `--extra-profiler-run`. This patch changes these
argument parsing code to make sure that we parse the profiler arguments for
both cases properly.

Note that `args.extra_profiler_run` might be true for apps that we don't
support. That's why we also check that the `app` is either one of the
GECKO_PROFILER_APPS or TRACE_APPS.

Differential Revision: https://phabricator.services.mozilla.com/D199650

UltraBlame original commit: 5b884354e9e1ce81092bac57532f8768b386c075