Bug 1878575 - Remove build/rust/bindgen-0.64. r=nical
The only use of it is nss-gk-api which can work with 0.69 as per its
dependency definition.

It seems cargo is removing this package, then updating bindgen (even
though it shouldn't), but this works around the issue and is harmless.

Differential Revision: https://phabricator.services.mozilla.com/D200676

UltraBlame original commit: d67e058935bae8bfad9fb5868b225926a25cb64a