Bug 1878180 - Catch more rust errors in logs. r=releng-reviewers,aryx,jcristau
Not all rust errors start with "error[Exxx]:", as shown by the two other
types of errors we catch that start with error:, but simplifying to
"error:" would be too broad, catching messages about std::error::Error.
"error: " might work, but feels risky (and it would be redundant with
the matchers for other compiler errors).

All the errors from the rust compiler appear at the beginning of the
line, though, so it feels a little safer to match on "error: " being at
the beginning (+ the optional [Exxx]), and it feels rather safe to
assume that if a line begins with error:, it's an error. This obviously
would catch other messages that start with error:, and maybe it should
be used more generally, but I want to see if this causes unexpected
problems first. (which is why I also went with fixing bug 1877891
separately of this)

Differential Revision: https://phabricator.services.mozilla.com/D200396

UltraBlame original commit: ba97fc77e52606be48d1dfe61aba1c5695afd652