Bug 1878784 - Allow `about:asrouter` to use `script-src`. r=ckerschb,mconley
There are three exceptions made for `about:newtab`; this updates the
first exception.  It does not appear that the `about:asrouter` CSP
includes a web scheme, so the second exception in the file is not
needed; nor does it appear that `about:asrouter` require
`unsafe-inline`, so the third exception in the file is not needed.

Differential Revision: https://phabricator.services.mozilla.com/D200733

UltraBlame original commit: d4cbb150e4c1c845c6984699902a3383f45cc230