Bug 1873394 - part3 : record telemetry for the failure to recreate the state machine. r=jolin,chutten
When enaling the media engine playback, we would always start with using
the external state machine. Then fallback to the normal state machine if
the conditions are not met, eg. only enable encrypted media playback, or
the codec types/key system types which the media engine doesn't support.

This kind of transistion should be smooth and always succeed. If not,
then it would probably cause bad effect on the user experience, so we
add the probes to measure the error rate.

Differential Revision: https://phabricator.services.mozilla.com/D198242

UltraBlame original commit: 89318af79fc6dae17b527682cc9b15062f876e7d