Bug 1877349 - Explicitly install the same dependencies that are installed during a Linux bootstrap in the `debian-build` Dockerfile as well r=firefox-build-system-reviewers,sergesanspaille,glandium
python-distutils was installing m4 implicitly. Since it was removed as a
part of the distutils removal work, it broke the build in CI. Let's
explicitly install all the dependencies we expect to avoid something
like this in the future.

https://searchfox.org/mozilla-central/rev/2a867dd1ab015c3ef24b774a57709fb3b3dc4961/python/mozboot/mozboot/linux_common.py#76-84

Differential Revision: https://phabricator.services.mozilla.com/D200500

UltraBlame original commit: 247b0488d16db53cdfe2a726e181215bd19789d6