Bug 1875409 - Remove the buggy SIGCHLD handler from the fork server. r=gsvelto
Signal handlers need to preserve `errno` in order to not change the
behavior of the code they interrupted; this one wasn't doing that, and
TSan complained.  In this case, the intended effect of the handler is
essentially the same as setting the signal disposition to `SIG_IGN`, so
the handler can just be removed.

Differential Revision: https://phabricator.services.mozilla.com/D201191

UltraBlame original commit: 23d502ed06280881577c5f9df617026ee30500f1