Bug 1877580 [wpt PR 44303] - [EditContext] Ensure textupdate fires for backspace/delete in <canvas>, a=testonly
Automatic update from web-platform-tests
[EditContext] Ensure textupdate fires for backspace/delete in <canvas>

When the active EditContext's associated element is a <canvas>, we
fail to fire the textudpate event when the user hits backspace/delete.

The reason is that for these keystrokes, EnabledInEditableText is
used to determine whether the command is enabled, and this includes
a check that the selection is contained in the focus. For <canvas>
this will never be the case because <canvas> cannot contain selection.

Fix this by always treating the command as enabled if there is an
active EditContext, since an EditContext should always have the
opportunity to handle these commands.

Update some existing text input and text deletion tests to run for
<canvas> in addition to <div>.

Bug: 1517103, 999184
Change-Id: Id8d4509ed6b22a4516d9bfa2c4a05b234eea78de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5251988
Reviewed-by: Anupam Snigdha <snianumicrosoft.com>
Reviewed-by: Koji Ishii <kojiichromium.org>
Commit-Queue: Dan Clark <daniecmicrosoft.com>
Cr-Commit-Position: refs/heads/main{#1257455}

--

wpt-commits: 12a51d92951f2858e1ad74a4412279e645e67a61
wpt-pr: 44303

UltraBlame original commit: 812a65af6d1d8bf09682c371ba3bf6c2ab85999f