Bug 1878506 - Add pref to make native HTTPS resolution exit early in automation r=necko-reviewers,kershaw
The purpose of this pref is to avoid breaking any tests that don't expect a
HTTPS record to be present. For example, if you're loading http://domain.com
in a test, and don't expect a HTTPS upgrade to happen, if that domain suddenly
adds a HTTPS record we might end up upgrading to HTTPS even in automation.

This pref does an early return with NS_ERROR_UNKNOWN_HOST just before doing
a call to the system API. This means we're still waiting in the DNS queue to
resolve the domain (simulating the same waiting times we might see
on users' computers), but we never actually use native DNS to resolve a HTTPS
record in automation.

Differential Revision: https://phabricator.services.mozilla.com/D200785

UltraBlame original commit: 9737908480c63eeb1bf2e74982b9a88a1bedb7e3