Bug 1879197 - Part 5: Port "Track ports that are going to be merged", r=ipc-reviewers,mccr8
This is a port of a chromium change originally reviewed at
https://chromium-review.googlesource.com/c/chromium/src/+/3289065

Original Author: Stephen Röttger <sroettgergoogle.com>

Original Description:

If a malicious node can leak a port name, they can initiate a port merge
with it and potentially inject messages.
To prevent this, we mark ports sent out for merges and verifiy on an
incoming merge request that the local port is expected to be used.
Ideally, we would track the peer node that should do the merge, but this
change should make exploitation already very unlikely since you'd need
to leak a port name and race the real node that will do the merge
itself.

Depends on D201108

Differential Revision: https://phabricator.services.mozilla.com/D201109

UltraBlame original commit: b7be8773d704960b06e34ed832897e93bc9484de