Bug 1880558: Allow PnP windows to enter macOS native fullscreen. r=pip-reviewers,mconley
This makes it possible, but not forced, for PnP windows to enter macOS
native fullscreen. This change makes the decision of whether to use
native or "emulated" fullscreen dependent on the setting of the
`full-screen-api.macos-native-full-screen` pref, same as top-level
windows. This is both more consistent and a bit strange in that it's
possible to make a YouTube fullscreen, then open a PnP window (which
will appear in the desktop space) and then make the PnP window
fullscreen in  *another* fullscreen space.

Although this is unusual to have multiple fullscreen spaces for the same
application, it is consistent with the macOS fullscreen experience. If
we don't want this behavior, we just need to decide what other kind of
behavior we'd prefer in this circumstance, such as:

* Make the PnP fullscreen replace the existing fullscreen window
space, effectively hiding the original window.
* Make the PnP window close instead of going to fullscreen.
* Remove the fullscreen option from the PnP player if the source window
is already in fullscreen.

Whether we want any of these changes or none of these changes, the
change in this patch is more correct when the pref
`full-screen-api.macos-native-full-screen` is set to true, since we
don't want to have a mix of native and emulated fullscreen behaviors.

Differential Revision: https://phabricator.services.mozilla.com/D202029

UltraBlame original commit: 0465723cfb4ca46ccf93d696eb564093a1d9f2e6