Bug 1881019 part 1 - Optimize TenuringTracer::traverse a bit. r=jonco
On Speedometer 3, about 48% of the values aren't GC things so check for
this first. Of the remaining values, about 47% are tenured so check for that
next.

This reduces the number of branches we have to take and improves some Speedometer 3
sub tests by a few percent.

Differential Revision: https://phabricator.services.mozilla.com/D202220

UltraBlame original commit: 1193e9ba23df1a6480ce610bbf97c1d343a53451