Bug 1861533: Allow checking multiple ArgumentsObject flags, r=jandem
The baseline tier sometimes checks whether any of a bit-set of flags are
present on an ArgumentsObject, so PBL needs the ability to do the same.

As a bonus, factoring this pattern to a separate function allows
simplifying the implementation of the existing getters for flags.

Depends on D192030

Differential Revision: https://phabricator.services.mozilla.com/D192031

UltraBlame original commit: 6a531229c73c90e50e9dd4bdc0c5a7f313206716