Bug 1878993 - [1/1] Inform Windows about different process types  r=win-reviewers,yjuglaret
Windows has an ill-documented subsystem known as the Application Launch
Prefetcher, which apparently preloads DLLs for a binary based on what
DLLs that binary loaded last time.

For Firefox, that's not a great heuristic; we relaunch ourselves as a
subprocess with potentially completely different DLL settings. We're not
the only such application, though, and Windows _does_ have a way to
signal the ALPF to load different DLL sets depending on launch context.

Differential Revision: https://phabricator.services.mozilla.com/D202275

UltraBlame original commit: dccf526bcc0afbf9f7abd33abff453ced5a76335