Bug 1879417 - part2 : resolve pending input promise when sample is popped. r=jolin
The problem is discovered when playing the file `red-46x48.mp4` in
`test_playback_rate_playpause.html`. That file only contains ONE frame
which is 1 second long.

There is an issue within the media engine, which The won't return an
output frame to us, unless we tell it the stream is ended.

Therefore, after sending the only frame to the media engine, we need to

(1) keep the media format reader to keep requesting data by resolving
the pending promise when the input has been popped so that the format
reader can read EOS.

(2) send the ended event to the media engine in order to get the first
frame.

Differential Revision: https://phabricator.services.mozilla.com/D202336

UltraBlame original commit: f1772dfe55a9095105410a44b5ae35e80e3f3ca6