Bug 1879341 - Part 2: Provide a way for conditions to be checked when an event is registered - r=scunnane
To allow the use of strings in Remote Settings to add conditions to the events, we pass expressions
that will be evaluated by the FilterExpression component. We pass the event details and some of
the params of the function that might be useful in the future.

Differential Revision: https://phabricator.services.mozilla.com/D202038

UltraBlame original commit: e4f259bc39a8dc031c551cce1278e5518eb429db