Bug 1881989 - Make `AutoDeleteRangesHandler::ExtendOrShrinkRangeToDelete` handle it with the closest editable ancestor block or inline editing host r=m_kato
It's currently handling its job with the closest ancestor block which may be
non-editable and editing host which is either inline or block.  However, the
closest block is required for check whether the range won't be extended outside
the closest block of the common ancestor of the range and the range is
guaranteed that they are in an editing host.  Therefore, it's not required if
it's outside the editing host.  So, comparisons which check whether a node is
either/neither editing host or/nor ancestor block can get same result with
comparing with the closest one of the editing host or the closest editable
block.

Differential Revision: https://phabricator.services.mozilla.com/D202697

UltraBlame original commit: c2895c7db175458e315a65fd8d5e9385e60a76d3