Bug 1879225: Refactor IsFoo methods r=dminor
In [this commit](https://chromium-review.googlesource.com/c/v8/v8/+/4727690), V8 changed its IsFoo predicates from methods into free functions as part of the Tagged<T> work.

At first I was simply going to move the current implementation, but as I was writing a comment to explain why `bool IsByteArray(Object obj) { return true; }` was not as dumb as it looked, I decided that we could do better.

V8's ByteArrays are managed by the GC, and the typecheck can be done by looking at the map (shape). Our implementation of ByteArray is a length-prefixed array allocated by malloc, so we can't check the type the same way, but in debug builds we can stick a magic number in the header and validate that it hasn't been messed with.

We can use a dedicated MagicValue for the exception we throw when interrupting (which maps closely to V8's approach).

Differential Revision: https://phabricator.services.mozilla.com/D201916

UltraBlame original commit: 8efe813ab556886d66e0d1b5a25dc865f30d24c9