Bug 1882425: Use a UniquePtr to manage Document::mSubDocuments. r=dom-core,edgar
This patch doesn't change behavior.

As noted in the code-comment: ideally we should fully-modernize to nsTHashMap
or similar, but that's a larger project and will require a bit of care to be
sure we're preserving the semantics that this struct imposes on itself right
now (e.g. subdocuments clearing their parent-document pointer, as part of
removal from the map). For now, I'm just taking the incremental/trivial win of
using UniquePtr to remove explicit `new`/`delete` statements.

Differential Revision: https://phabricator.services.mozilla.com/D202904

UltraBlame original commit: da466faa7aee60bdeeca32b0767a1f0b9b964272