Bug 1880078 - A previously opened tab is shown in Address Bar after restarting the browser and closing the tab. r=jteow
Some consumer is passing in integer as expected, most pass a numeric string
that is the result of a getAttribute. This causes a mismatch as that id is used
as key in a Map.
While it would be possible to fix all the consumers, the mistake seems so
common and trivial, that is worth just adapting the API to accept both.

Differential Revision: https://phabricator.services.mozilla.com/D203019

UltraBlame original commit: e594b8a45404bb3c1e37221eeea1b166cd5cde47