Bug 1835509. If we encounter a bad LZW block in a gif frame still report success if we have decoded frames before this successfully. r=aosmond
This is similar to this code in image/Decoder.cpp

https://searchfox.org/mozilla-central/rev/ae292ebba6074601b33fa983dd4e01ce6a1ec4ac/image/Decoder.cpp#257

We won't display the frames after the bad frame, but that seems reasonable, and an improvement from current. If we get more reports of this we can look into just skipping the bad frame instead of truncating before the bad frame.

Differential Revision: https://phabricator.services.mozilla.com/D179495

UltraBlame original commit: 7ebd26a0e71e71dfab98d9026f69ed29c2289878