Bug 1882828 - Ensure hard-coded batched texture uploads enabled state does not get overridden by pref. r=gfx-reviewers,lsalzman
Batched texture uploads are supposed to be hard-coded as enabled or
disabled on certain platforms, and on other platforms controllable via
a pref. However, a previous refactoring made it so that the hard-coded
value is always being overridden by the pref (by checking whether the
hard-coded value is set in the wrong place). As a consequence, batched
texture uploads have been disabled on Mali-G devices, resulting in
poor performance.

This patch moves the check to the correct place, meaning the
hard-coded value is respected once again.

Differential Revision: https://phabricator.services.mozilla.com/D203145

UltraBlame original commit: aaed3db903c2afdf8b5c6df0d93dac6cddf72a37