Bug 1882817 - Convert reftests for stretching by scaling to internal WPT tests. r=emilio
These tests were introduced in bug 414277 to verify stretching fallback
using a scale transform. The tests assume that the default fonts don't
provide a way to stretch the operators in the tests (e.g. triple arrows
and quadruple integrals) so that the scale stretching fallback. Moreover
it assumes that the operators stretch perfectly to match the target
size.

These assumptions may be incorrect and are difficult to ensure, which
is likely why one test fails on mac (which as STIX pre-installed). For
now, we keep the test as is but just make sure the preference
`mathml.scale_stretchy_operators.enabled` is enabled.

Differential Revision: https://phabricator.services.mozilla.com/D203439

UltraBlame original commit: a1e252ada9c21f13326e2cd203ba321de365e9bd