Bug 1882817 - Convert scriptlevel-movablelimits-1.html reftest to internal WPT test. r=emilio
This test verifies the MathML3 rule that the `accentunder`/`accent`
attributes are ignored when an `munderover` element (and similar
scripted elements) is laid out as an `msubsup` element (i.e. when the
base is a `movablelimits` operator in `displaystyle="false"`).

More specifically, setting `accentunder`/`accent` to `true` normally
prevents the `scriptlevel` to be incremented in an `munderover` element
(and similar scripted elements) but this test verifies that this rule
does not hold when such an element is laid out as an `msubsup` element.

In MathML Core, the CSS-based mechanism is slightly different: the
scripts of `munderover` (and similar scripted elements) always have
`math-depth: add(1)` in the UA sheet but `font-size: inherit` if the
corresponding `accentunder`/`accent` attribute is set to true. This
means that by default, the `accentunder`/`accent` scripts are not drawn
smaller even when laid out as an `msubsup`, contrary to MathML3.

Consequently, we keep this test as internal for now.

Differential Revision: https://phabricator.services.mozilla.com/D203577

UltraBlame original commit: 7e30466613303e4d2be13916486d2bb94cb1782b