Bug 1883657 - Remove isEarlyStartup flag from backgroundUpdate r=nalexander
This flag was only used to set the setPref prefs on the user branch. Bug
1875404 updated the setPref annotation to specify the branch for each pref, so
the isEarlyStartup flag was not doing anything for this feature.

Additionally, bug 1875404 mistakenly updated the pref to be set on the default
branch, but isEarlyStartup setPrefs are set on the user branch. The pref has
been updated to be set on the user branch.

Differential Revision: https://phabricator.services.mozilla.com/D203627

UltraBlame original commit: 54c01d6de67c82cfdc0be847b8d83e5f0a652bfd