Bug 1884268: Remove FirstSubsumedFrame::operator=(&&) r=jandem
There was a bug in this code that was caught and fixed in review, but we accidentally landed a previous version of the patch. Then it sat unnoticed for eight years because it was never called.

We could land the fixed version, but it's simpler to just delete it.

Differential Revision: https://phabricator.services.mozilla.com/D204132

UltraBlame original commit: 7ec159556387e97c13c7ad7e67152187553b02da