Bug 1877218 - Mention mach-try-perf in the selectors doc, and a warning to try chooser. r=ahal DONTBUILD
This patch adds a mention of the perf selector to the selectors documentation. It also adds a warning to the top of the Performance panel in the try chooser to warn users not to use it, and use the `./mach try perf` tool instead.

Differential Revision: https://phabricator.services.mozilla.com/D203800

UltraBlame original commit: e7bb5a728361f4b05cdc970a020bbe0b89e1a10b