Bug 1884232 - Decrease the bottom padding of the urlbar's dismissal acknowledgment tip so it's balanced with its top padding. r=dao
The problem is that the bottom padding of `.urlbarView-results` is right up
against the bottom padding of the tip when the tip is the last result, making
the effective visual padding 4px + 6px = 10px, whereas the tip's top padding is
only 6px. We just need to subtract the `.urlbarView-results` padding from the
tip's padding when the tip is the last child.

Differential Revision: https://phabricator.services.mozilla.com/D203966

UltraBlame original commit: 59fc9dd3d84303161f522bd7735119fb86649186