Bug 1883902 [wpt PR 44951] - Fix timeout in svg-mouse-listener a11y test, a=testonly
Automatic update from web-platform-tests
Fix timeout in svg-mouse-listener a11y test

The recommended approach to make a crash test asynchronous is to use
reftest-wait. Unfortunately, test-driver does not work with ref-tests.
The fix is to pull the test out of crashtests and author as a promise
test.  This is the only a11y crash test that uses test-driver, and
thus, the only one that is affected.

The test needs to be asynchronous to avoid a memory leak. The reason is
due to the fact that we have armed an event listener that is never
triggered unless we wait. A crash-test only waits until onload is complete.

Bug: 328403023
Change-Id: If33410bd498995b8216e805c8edc27c093e3db1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5348198
Reviewed-by: Aaron Leventhal <aleventhalchromium.org>
Commit-Queue: Kevin Ellis <keverschromium.org>
Cr-Commit-Position: refs/heads/main{#1269156}

--

wpt-commits: fcb1f27c332516c08aed3c0d668cf267f7671c28
wpt-pr: 44951

UltraBlame original commit: 3d08dc2ef6034a4b0d2353e4e0521bee32b83514