[components] Closes https://github.com/mozilla-mobile/android-components/issues/904: Refactor SessionStorage around idea of snapshots
This patch introduces an idea of a SessionsSnapshot, which is a representation of
SessionManager's state that's suitable for persistance. SessionStorage's APIs read/write
APIs are changed to operate over the snapshot.

A public `restore` method was added to SessionManager, along with onSessionsRestored observer method,
which allow restoring a SessionsSnapshot.

UltraBlame original commit: cd4c229ff441d578ae2bdbe19a2f871c773a451f