[components] Remove references to the missing timezone field
In bug 1518165 it was decided that the timezone information
in `start_time`, `end_time` and the date header is sufficient
for analyses and that we won't need an additional timezone
field. This gets rid of the vestigial comments.

UltraBlame original commit: 26ef4bd3847bba76be4351bdcf4a1f0267e25cbc