[components] Part 2: Move FirefoxSyncFeature into service-sync, rename as StorageSync
We will need to access this functionality from the SyncDispatcher implementation,
and so we need to move it out of a feature component. This code is not really
a "feature" in a sense of how our other features behave, either.

UltraBlame original commit: 8586180ae1a003a5b622b57cccf240971cc9e9e0