[components] Fix blank custom tabs following an NPE in sample-browser
Fixes https://github.com/mozilla-mobile/android-components/issues/2673.

This PR protects against the initial `NullPointerException` ever happening.

This is a rare case, and we do not have anything to go on at this point, so we fallback to a new intent, and the user is routed to the home activity.

UltraBlame original commit: 4ac9ac759773f4388b6e9a5e45da871ade4b17c0