[components] Closes https://github.com/mozilla-mobile/android-components/issues/2862: Make VisitType a requirement for HistoryTrackingDelegate's onVisited method
This shifts responsibility of constructing a VisitType for any given visit on the
caller of HistoryTrackingDelegate implementation, which is normally the engine.

This is generally what we want: engine is normally best suited to determine a visit type
for a URI that's currently being opened.

UltraBlame original commit: f87039edc92bc7bf0d17b8cf6b57289f62656453