[components] Add check to ensure Glean is only initialized on the main process.
Trying to initialize Glean on a process other than the main process results in an error when Glean tries to use WorkManager because WorkManager is only initialized on the main process.  This checks whether initialize is being called by the main process and returns if it isn't.

UltraBlame original commit: a2ad5d542e13971d6b97467718936fe7183815ab