[components] Fix for intermittent test failures in Glean
Due to upgrading WorkManager dependency, testing of WorkManager now works so all of the tests depending on our workaround in `triggerWorkManager` needed a fix to prevent intermittent failures.  This required us to use the WorkManager TestDriver to trigger the conditions and now we test WorkManager as it is supposed to be tested.

UltraBlame original commit: 95592c3284115e49dfb774e1fd538dcebce17c58