[components] Always set 'wasMigrated' to false in the Glean SDK
This is only useful as a safety measure in case the data migration plan goes
wrong and we're forced to downgrade the version Glean SDK consumers are using
from glean-core to glean-ac. In that case, this code will run and will reset
the status of the migration. When attempting the upgrade again, glean-core
will perform the migration one more time.

UltraBlame original commit: 52a77465f60e72f9cfef10ec8704b3dacb90f589