[components] Merge https://github.com/mozilla-mobile/android-components/pull/4958
4958: Password sync support r=pocmo a=grigoryk

This does a few things in order to enable password syncing.

Our underlying storage for passwords is encrypted at rest. This means that we need to keep the passwords storage layer unlocked for the duration of a sync, which is what this PR achieves.

In order to keep password storage layer unlocked, we need to be able to access the encryption key from the worker. To achieve that, worker is now using the new `KeySharedPreferences` API to obtain a `passwords_key`.

`KeySharedPreferences` was added to make it easier to transparently store key/values pairs such that `value` is encrypted at rest, protected by a key which is stored in AndroidKeyStore. Due to issues with the AndroidKeyStore for API versions prior to 23, on those older API levels `KeySharedPreferences` simply delegates to SharedPreferences, without any encryption.

`KeySharedPreferences` supports an API upgrade scenario when going from pre-API23 to post-API23, but not the other way.



Co-authored-by: Emily Kager <ekagermozilla.com>
Co-authored-by: Grisha Kruglov <gkruglovmozilla.com>

UltraBlame original commit: a07d06e5201d9c4d5db035dbd5390837d0f647ff