[components] Merge https://github.com/mozilla-mobile/android-components/pull/5065
5065: Closes https://github.com/mozilla-mobile/android-components/issues/4935: Tracking protection state not updated on browser session r=pocmo,Amejia481 a=csadilek

The cause of this problem is a bit of a gap in our design. When a `GeckoEngineSession` is created we enable (or disable) tracking protection based on settings, and also call `notifyObservers` to make sure the corresponding browser session (`trackerBlockingEnabled` field) is updated. However, when the `GeckoEngineSession` is created no browser session is linked to it yet and therefore there's no observer. The notification is simply lost and the browser session is never updated.

We looked into a few ways to work around this e.g. setting the field in [onLocationChange](https://github.com/mozilla-mobile/android-components/blob/master/components/browser/session/src/main/java/mozilla/components/browser/session/engine/EngineObserver.kt#L41) instead, as we do for `title`, `url`, `webAppManifest`. This would be a bit of a hack though as we'd be doing it way too often, or in an unrelated method / event.

So, how about this idea instead: We introduce a new `notifyAtLeastOneObserver` (or similar) method which queues up the notification if there's no observer and sends it to the first observer that's registered. Wdyt?

pocmo Amejia481 


Co-authored-by: Christian Sadilek <christian.sadilekgmail.com>

UltraBlame original commit: 818fb03c7791e2bb439b11154f598a03c3bf2dce